+++ src/main/scala/edu/ie3/simona/agent/participant/ParticipantAgent.scala
@@ -16,8 +16,8 @@ import edu.ie3.simona.exceptions.CriticalFailureException
 import edu.ie3.simona.model.participant.ParticipantModel.AdditionalFactoryData
 import edu.ie3.simona.model.participant.ParticipantModelShell
 import edu.ie3.simona.ontology.messages.SchedulerMessage.Completion
-import edu.ie3.simona.ontology.messages.flex.FlexibilityMessage._
-import edu.ie3.simona.ontology.messages.services.ServiceMessage
+import edu.ie3.simona.ontology.messages.flex.FlexibilityMessage.*
+import edu.ie3.simona.ontology.messages.ServiceMessage
 import edu.ie3.simona.ontology.messages.{Activation, SchedulerMessage}
 import edu.ie3.simona.service.Data
 import edu.ie3.simona.service.Data.{PrimaryData, PrimaryDataExtra}
@@ -52,13 +52,13 @@ object ParticipantAgent {
   /** Messages that are sent by services as responses to registration requests.
     */
   sealed trait RegistrationResponseMessage extends Request {
-    val serviceRef: ActorRef[? >: ServiceMessage]
+    val serviceRef: ActorRef[ServiceMessage]
   }
 
   /** Message confirming a successful registration with a secondary service.
     */
   final case class RegistrationSuccessfulMessage(
-      override val serviceRef: ActorRef[? >: ServiceMessage],
+      override val serviceRef: ActorRef[ServiceMessage],
       firstDataTick: Long,
       additionalData: Option[AdditionalFactoryData] = None,
   ) extends RegistrationResponseMessage
@@ -73,7 +73,7 @@ object ParticipantAgent {
     *   The type of primary data to be received.
     */
   final case class PrimaryRegistrationSuccessfulMessage[P <: PrimaryData](
-      override val serviceRef: ActorRef[? >: ServiceMessage],
+      override val serviceRef: ActorRef[ServiceMessage],
       firstDataTick: Long,
       primaryDataExtra: PrimaryDataExtra[P],
   ) extends RegistrationResponseMessage
@@ -81,7 +81,7 @@ object ParticipantAgent {
   /** Message announcing a failed registration.
     */
   final case class RegistrationFailedMessage(
-      override val serviceRef: ActorRef[? >: ServiceMessage]
+      override val serviceRef: ActorRef[ServiceMessage]
   ) extends RegistrationResponseMessage
 
   /** Data provision messages sent by data services.
@@ -94,7 +94,7 @@ object ParticipantAgent {
 
     /** The sending service actor ref.
       */
-    val serviceRef: ActorRef[? >: ServiceMessage]
+    val serviceRef: ActorRef[ServiceMessage]
 
     /** Next tick at which data could arrive. If None, no data is expected for
       * the rest of the simulation.
@@ -111,7 +111,7 @@ object ParticipantAgent {
     */
   final case class DataProvision[D <: Data](
       override val tick: Long,
-      override val serviceRef: ActorRef[? >: ServiceMessage],
+      override val serviceRef: ActorRef[ServiceMessage],
       data: D,
       override val nextDataTick: Option[Long],
   ) extends DataInputMessage
@@ -123,7 +123,7 @@ object ParticipantAgent {
     */
   final case class NoDataProvision(
       override val tick: Long,
-      override val serviceRef: ActorRef[? >: ServiceMessage],
+      override val serviceRef: ActorRef[ServiceMessage],
       override val nextDataTick: Option[Long],
   ) extends DataInputMessage