+++ src/test/scala/edu/ie3/simona/agent/em/EmAgentIT.scala
@@ -18,7 +18,7 @@ import edu.ie3.simona.agent.participant.ParticipantAgentInit.{
   ParticipantRefs,
   SimulationParameters,
 }
-import edu.ie3.simona.config.RuntimeConfig._
+import edu.ie3.simona.config.RuntimeConfig.*
 import edu.ie3.simona.event.ResultEvent
 import edu.ie3.simona.event.ResultEvent.ParticipantResultEvent
 import edu.ie3.simona.event.notifier.NotifierConfig
@@ -26,22 +26,24 @@ import edu.ie3.simona.ontology.messages.SchedulerMessage.{
   Completion,
   ScheduleActivation,
 }
-import edu.ie3.simona.ontology.messages.services.ServiceMessage
-import edu.ie3.simona.ontology.messages.services.ServiceMessage.PrimaryServiceRegistrationMessage
-import edu.ie3.simona.ontology.messages.services.WeatherMessage.{
-  RegisterForWeatherMessage,
-  WeatherData,
+import edu.ie3.simona.ontology.messages.ServiceMessage.{
+  PrimaryServiceRegistrationMessage,
+  SecondaryServiceRegistrationMessage,
 }
 import edu.ie3.simona.ontology.messages.{Activation, SchedulerMessage}
 import edu.ie3.simona.scheduler.ScheduleLock
+import edu.ie3.simona.service.Data.SecondaryData.WeatherData
 import edu.ie3.simona.service.ServiceType
+import edu.ie3.simona.service.primary.PrimaryServiceProxy
+import edu.ie3.simona.service.weather.WeatherService.Coordinate
+import edu.ie3.simona.service.weather.{WeatherService, WeatherServiceSpec}
 import edu.ie3.simona.test.common.TestSpawnerTyped
 import edu.ie3.simona.test.common.input.EmInputTestData
-import edu.ie3.simona.util.SimonaConstants.{INIT_SIM_TICK, PRE_INIT_TICK}
 import edu.ie3.simona.test.matchers.QuantityMatchers
+import edu.ie3.simona.util.SimonaConstants.{INIT_SIM_TICK, PRE_INIT_TICK}
 import edu.ie3.simona.util.TickUtil.TickLong
 import edu.ie3.util.TimeUtil
-import edu.ie3.util.quantities.QuantityUtils._
+import edu.ie3.util.quantities.QuantityUtils.*
 import edu.ie3.util.scala.quantities.WattsPerSquareMeter
 import org.apache.pekko.actor.testkit.typed.scaladsl.{
   ScalaTestWithActorTestKit,
@@ -104,8 +106,8 @@ class EmAgentIT
         val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
         val resultListener = TestProbe[ResultEvent]("ResultListener")
         val primaryServiceProxy =
-          TestProbe[ServiceMessage]("PrimaryServiceProxy")
-        val weatherService = TestProbe[ServiceMessage]("WeatherService")
+          TestProbe[PrimaryServiceProxy.Message]("PrimaryServiceProxy")
+        val weatherService = TestProbe[WeatherService.Message]("WeatherService")
         val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
         given ParticipantRefs = ParticipantRefs(
@@ -201,10 +203,12 @@ class EmAgentIT
 
         // deal with weather service registration
         weatherService.expectMessage(
-          RegisterForWeatherMessage(
+          SecondaryServiceRegistrationMessage(
             pvAgent,
-            pvInput.getNode.getGeoPosition.getY,
-            pvInput.getNode.getGeoPosition.getX,
+            Coordinate(
+              pvInput.getNode.getGeoPosition.getY,
+              pvInput.getNode.getGeoPosition.getX,
+            ),
           )
         )
 
@@ -337,8 +341,8 @@ class EmAgentIT
         val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
         val resultListener = TestProbe[ResultEvent]("ResultListener")
         val primaryServiceProxy =
-          TestProbe[ServiceMessage]("PrimaryServiceProxy")
-        val weatherService = TestProbe[ServiceMessage]("WeatherService")
+          TestProbe[PrimaryServiceProxy.Message]("PrimaryServiceProxy")
+        val weatherService = TestProbe[WeatherService.Message]("WeatherService")
         val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
         given ParticipantRefs = ParticipantRefs(
@@ -434,10 +438,12 @@ class EmAgentIT
 
         // deal with weather service registration
         weatherService.expectMessage(
-          RegisterForWeatherMessage(
+          SecondaryServiceRegistrationMessage(
             pvAgent,
-            pvInput.getNode.getGeoPosition.getY,
-            pvInput.getNode.getGeoPosition.getX,
+            Coordinate(
+              pvInput.getNode.getGeoPosition.getY,
+              pvInput.getNode.getGeoPosition.getX,
+            ),
           )
         )
 
@@ -448,10 +454,12 @@ class EmAgentIT
 
         // deal with weather service registration
         weatherService.expectMessage(
-          RegisterForWeatherMessage(
+          SecondaryServiceRegistrationMessage(
             hpAgent,
-            adaptedHpInputModel.getNode.getGeoPosition.getY,
-            adaptedHpInputModel.getNode.getGeoPosition.getX,
+            Coordinate(
+              adaptedHpInputModel.getNode.getGeoPosition.getY,
+              adaptedHpInputModel.getNode.getGeoPosition.getX,
+            ),
           )
         )
 
@@ -636,8 +644,8 @@ class EmAgentIT
         val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
         val resultListener = TestProbe[ResultEvent]("ResultListener")
         val primaryServiceProxy =
-          TestProbe[ServiceMessage]("PrimaryServiceProxy")
-        val weatherService = TestProbe[ServiceMessage]("WeatherService")
+          TestProbe[PrimaryServiceProxy.Message]("PrimaryServiceProxy")
+        val weatherService = TestProbe[WeatherService.Message]("WeatherService")
         val scheduler = TestProbe[SchedulerMessage]("Scheduler")
 
         given ParticipantRefs = ParticipantRefs(
@@ -725,10 +733,12 @@ class EmAgentIT
 
         // deal with weather service registration
         weatherService.expectMessage(
-          RegisterForWeatherMessage(
+          SecondaryServiceRegistrationMessage(
             pvAgent,
-            pvInputLimitedOperationTime.getNode.getGeoPosition.getY,
-            pvInputLimitedOperationTime.getNode.getGeoPosition.getX,
+            Coordinate(
+              pvInputLimitedOperationTime.getNode.getGeoPosition.getY,
+              pvInputLimitedOperationTime.getNode.getGeoPosition.getX,
+            ),
           )
         )