+++ src/test/scala/edu/ie3/simona/agent/grid/ThermalGridIT.scala
@@ -26,15 +26,21 @@ import edu.ie3.simona.ontology.messages.SchedulerMessage.{
   Completion,
   ScheduleActivation,
 }
-import edu.ie3.simona.ontology.messages.services.ServiceMessage
-import edu.ie3.simona.ontology.messages.services.ServiceMessage.PrimaryServiceRegistrationMessage
-import edu.ie3.simona.ontology.messages.services.WeatherMessage.{
-  RegisterForWeatherMessage,
-  WeatherData,
+import edu.ie3.simona.ontology.messages.ServiceMessage.{
+  PrimaryServiceRegistrationMessage,
+  SecondaryServiceRegistrationMessage,
 }
-import edu.ie3.simona.ontology.messages.{Activation, SchedulerMessage}
+import edu.ie3.simona.ontology.messages.{
+  Activation,
+  SchedulerMessage,
+  ServiceMessage,
+}
 import edu.ie3.simona.scheduler.ScheduleLock
+import edu.ie3.simona.service.Data.SecondaryData.WeatherData
 import edu.ie3.simona.service.ServiceType
+import edu.ie3.simona.service.primary.PrimaryServiceProxy
+import edu.ie3.simona.service.weather.WeatherService.Coordinate
+import edu.ie3.simona.service.weather.WeatherService
 import edu.ie3.simona.test.common.input.EmInputTestData
 import edu.ie3.simona.test.common.{DefaultTestData, TestSpawnerTyped}
 import edu.ie3.simona.test.matchers.QuantityMatchers
@@ -110,9 +116,8 @@ class ThermalGridIT
       val resultListener = TestProbe[ResultEvent]("ResultListener")
       val scheduler: TestProbe[SchedulerMessage] = TestProbe("scheduler")
       val primaryServiceProxy =
-        TestProbe[ServiceMessage]("PrimaryServiceProxy")
-
-      val weatherService = TestProbe[ServiceMessage]("WeatherService")
+        TestProbe[PrimaryServiceProxy.Message]("PrimaryServiceProxy")
+      val weatherService = TestProbe[WeatherService.Message]("WeatherService")
 
       given ParticipantRefs = ParticipantRefs(
         gridAgent = gridAgent.ref,
@@ -158,10 +163,12 @@ class ThermalGridIT
       hpAgent ! RegistrationFailedMessage(primaryServiceProxy.ref)
 
       weatherService.expectMessage(
-        RegisterForWeatherMessage(
+        SecondaryServiceRegistrationMessage(
           hpAgent,
-          typicalHpInputModel.getNode.getGeoPosition.getY,
-          typicalHpInputModel.getNode.getGeoPosition.getX,
+          Coordinate(
+            typicalHpInputModel.getNode.getGeoPosition.getY,
+            typicalHpInputModel.getNode.getGeoPosition.getX,
+          ),
         )
       )
 
@@ -731,9 +738,8 @@ class ThermalGridIT
       val resultListener: TestProbe[ResultEvent] = TestProbe("resultListener")
       val scheduler: TestProbe[SchedulerMessage] = TestProbe("scheduler")
       val primaryServiceProxy =
-        TestProbe[ServiceMessage]("PrimaryServiceProxy")
-
-      val weatherService = TestProbe[ServiceMessage]("WeatherService")
+        TestProbe[PrimaryServiceProxy.Message]("PrimaryServiceProxy")
+      val weatherService = TestProbe[WeatherService.Message]("WeatherService")
 
       given ParticipantRefs = ParticipantRefs(
         gridAgent = gridAgent.ref,
@@ -816,10 +822,12 @@ class ThermalGridIT
 
       // deal with weather service registration
       weatherService.expectMessage(
-        RegisterForWeatherMessage(
+        SecondaryServiceRegistrationMessage(
           pvAgent,
-          pvInput.getNode.getGeoPosition.getY,
-          pvInput.getNode.getGeoPosition.getX,
+          Coordinate(
+            pvInput.getNode.getGeoPosition.getY,
+            pvInput.getNode.getGeoPosition.getX,
+          ),
         )
       )
 
@@ -833,10 +841,12 @@ class ThermalGridIT
 
       // deal with weather service registration
       weatherService.expectMessage(
-        RegisterForWeatherMessage(
+        SecondaryServiceRegistrationMessage(
           hpAgent,
-          typicalHpInputModel.getNode.getGeoPosition.getY,
-          typicalHpInputModel.getNode.getGeoPosition.getX,
+          Coordinate(
+            typicalHpInputModel.getNode.getGeoPosition.getY,
+            typicalHpInputModel.getNode.getGeoPosition.getX,
+          ),
         )
       )