+++ src/test/scala/edu/ie3/simona/model/participant/evcs/EvcsModelIT.scala
@@ -16,7 +16,7 @@ import edu.ie3.simona.agent.participant.ParticipantAgentInit.{
 }
 import edu.ie3.simona.api.data.ev.ExtEvDataConnection
 import edu.ie3.simona.api.data.ev.model.EvModel
-import edu.ie3.simona.api.data.ev.ontology._
+import edu.ie3.simona.api.data.ev.ontology.*
 import edu.ie3.simona.api.data.ontology.ScheduleDataServiceMessage
 import edu.ie3.simona.config.RuntimeConfig.EvcsRuntimeConfig
 import edu.ie3.simona.event.ResultEvent
@@ -26,33 +26,37 @@ import edu.ie3.simona.ontology.messages.SchedulerMessage.{
   Completion,
   ScheduleActivation,
 }
-import edu.ie3.simona.ontology.messages.services.ServiceMessage
-import edu.ie3.simona.ontology.messages.services.ServiceMessage.{
+import edu.ie3.simona.ontology.messages.ServiceMessage.{
   Create,
   PrimaryServiceRegistrationMessage,
 }
-import edu.ie3.simona.ontology.messages.{Activation, SchedulerMessage}
+import edu.ie3.simona.ontology.messages.{
+  Activation,
+  SchedulerMessage,
+  ServiceMessage,
+}
 import edu.ie3.simona.scheduler.ScheduleLock
 import edu.ie3.simona.service.ServiceType
 import edu.ie3.simona.service.ev.ExtEvDataService
 import edu.ie3.simona.service.ev.ExtEvDataService.InitExtEvData
+import edu.ie3.simona.service.primary.PrimaryServiceProxy
 import edu.ie3.simona.test.common.input.EvcsInputTestData
 import edu.ie3.simona.test.common.{TestSpawnerTyped, UnitSpec}
 import edu.ie3.simona.util.SimonaConstants.{INIT_SIM_TICK, PRE_INIT_TICK}
 import edu.ie3.simona.util.TickUtil.TickLong
-import edu.ie3.util.quantities.QuantityUtils._
+import edu.ie3.util.quantities.QuantityUtils.*
 import org.apache.pekko.actor.testkit.typed.scaladsl.{
   ScalaTestWithActorTestKit,
   TestProbe,
 }
-import org.apache.pekko.actor.typed.scaladsl.adapter._
+import org.apache.pekko.actor.typed.scaladsl.adapter.*
 import squants.Each
 
 import java.time.ZonedDateTime
 import java.time.temporal.ChronoUnit
 import java.util.UUID
-import scala.jdk.CollectionConverters._
-import scala.jdk.OptionConverters._
+import scala.jdk.CollectionConverters.*
+import scala.jdk.OptionConverters.*
 
 /** Tests the combined functionality of
   * [[edu.ie3.simona.agent.participant.ParticipantAgent]] with an [[EvcsModel]]
@@ -91,7 +95,7 @@ class EvcsModelIT
       val gridAgent = TestProbe[GridAgent.Message]("GridAgent")
       val resultListener = TestProbe[ResultEvent]("ResultListener")
       val primaryServiceProxy =
-        TestProbe[ServiceMessage]("PrimaryServiceProxy")
+        TestProbe[PrimaryServiceProxy.Message]("PrimaryServiceProxy")
       val scheduler = TestProbe[SchedulerMessage]("Scheduler")
       val extSimAdapter = TestProbe[Any]("ExtSimAdapter")
 
@@ -101,15 +105,11 @@ class EvcsModelIT
         "ExtEvDataService",
       )
 
-      val adapterToExt = spawn(
-        ExtEvDataService.adapter(evService),
-        "ExtEvDataService-external-adapter",
-      )
-
       val extEvData = new ExtEvDataConnection()
+
       extEvData.setActorRefs(
-        adapterToExt.toClassic,
-        extSimAdapter.ref.toClassic,
+        evService,
+        extSimAdapter.ref,
       )
       val serviceKey =
         ScheduleLock.singleKey(TSpawner, scheduler.ref, PRE_INIT_TICK)
@@ -121,10 +121,9 @@ class EvcsModelIT
         serviceKey,
       )
 
-      val scheduleServiceMsg = scheduler.expectMessageType[ScheduleActivation]
-      scheduleServiceMsg.tick shouldBe INIT_SIM_TICK
-      scheduleServiceMsg.unlockKey shouldBe Some(serviceKey)
-      val serviceActivation = scheduleServiceMsg.actor
+      scheduler.expectMessage(
+        ScheduleActivation(evService, INIT_SIM_TICK, Some(serviceKey))
+      )
 
       /* Create ParticipantAgent with EvcsModel */
       given ParticipantRefs = ParticipantRefs(
@@ -156,8 +155,8 @@ class EvcsModelIT
 
       /* INIT */
 
-      serviceActivation ! Activation(INIT_SIM_TICK)
-      scheduler.expectMessage(Completion(serviceActivation))
+      evService ! Activation(INIT_SIM_TICK)
+      scheduler.expectMessage(Completion(evService))
 
       evcsActivation ! Activation(INIT_SIM_TICK)
 
@@ -174,50 +173,44 @@ class EvcsModelIT
         Map.empty[UUID, java.util.List[EvModel]].asJava,
         Some(long2Long(0L)).toJava,
       )
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
       // waiting for ExtEvDataService
       scheduler.expectNoMessage()
 
-      serviceActivation ! Activation(INIT_SIM_TICK)
+      evService ! Activation(INIT_SIM_TICK)
 
       scheduler.receiveMessages(2) should contain allOf (
         Completion(evcsActivation, Some(0)),
-        Completion(serviceActivation, None)
+        Completion(evService, None)
       )
 
       /* TICK 0 */
 
       // Request prices (dummy implementation)
       extEvData.sendExtMsg(new RequestCurrentPrices())
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(0)
+      evService ! Activation(0)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideCurrentPrices(
         Map(evcsInputModel.getUuid -> double2Double(0.0)).asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       // Request free lots
       extEvData.sendExtMsg(new RequestEvcsFreeLots())
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(0)
+      evService ! Activation(0)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideEvcsFreeLots(
         // No EV connected
         Map(evcsInputModel.getUuid -> int2Integer(2)).asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       resultListener.expectNoMessage()
 
@@ -228,16 +221,14 @@ class EvcsModelIT
         ).asJava,
         Some(long2Long(9000)).toJava,
       )
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
       // waiting for ExtEvDataService
       scheduler.expectNoMessage()
 
-      serviceActivation ! Activation(0)
+      evService ! Activation(0)
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       evcsActivation ! Activation(0)
 
@@ -273,11 +264,9 @@ class EvcsModelIT
 
       // Request free lots
       extEvData.sendExtMsg(new RequestEvcsFreeLots())
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(1800)
+      evService ! Activation(1800)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideEvcsFreeLots(
         // evA and evB connected
@@ -285,7 +274,7 @@ class EvcsModelIT
         Map.empty[UUID, java.lang.Integer].asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       resultListener.expectNoMessage()
 
@@ -346,11 +335,9 @@ class EvcsModelIT
 
       // Request free lots
       extEvData.sendExtMsg(new RequestEvcsFreeLots())
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(9000)
+      evService ! Activation(9000)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideEvcsFreeLots(
         // evB connected
@@ -358,7 +345,7 @@ class EvcsModelIT
         Map(evcsInputModel.getUuid -> int2Integer(1)).asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       // Request departing EVs
       extEvData.sendExtMsg(
@@ -366,33 +353,29 @@ class EvcsModelIT
           Map(evcsInputModel.getUuid -> List(evA.getUuid).asJava).asJava
         )
       )
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(9000)
+      evService ! Activation(9000)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideDepartingEvs(
         List[EvModel](evA.copyWith(10.0.asKiloWattHour)).asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       // Send (empty) arrivals in order to update next tick
       extEvData.provideArrivingEvs(
         Map.empty[UUID, java.util.List[EvModel]].asJava,
         Some(long2Long(10800)).toJava,
       )
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
       // waiting for ExtEvDataService
       scheduler.expectNoMessage()
 
-      serviceActivation ! Activation(9000)
+      evService ! Activation(9000)
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       evcsActivation ! Activation(9000)
 
@@ -405,18 +388,16 @@ class EvcsModelIT
 
       // Request free lots
       extEvData.sendExtMsg(new RequestEvcsFreeLots())
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(10800)
+      evService ! Activation(10800)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideEvcsFreeLots(
         // evB connected
         Map(evcsInputModel.getUuid -> int2Integer(1)).asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       resultListener.expectNoMessage()
 
@@ -427,16 +408,14 @@ class EvcsModelIT
         ).asJava,
         Some(long2Long(14400)).toJava,
       )
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
       // waiting for ExtEvDataService
       scheduler.expectNoMessage()
 
-      serviceActivation ! Activation(10800)
+      evService ! Activation(10800)
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       evcsActivation ! Activation(10800)
 
@@ -467,11 +446,9 @@ class EvcsModelIT
 
       // Request free lots
       extEvData.sendExtMsg(new RequestEvcsFreeLots())
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(12600)
+      evService ! Activation(12600)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideEvcsFreeLots(
         // evB and evC connected
@@ -479,7 +456,7 @@ class EvcsModelIT
         Map.empty[UUID, java.lang.Integer].asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       // EVCS activation
       evcsActivation ! Activation(12600)
@@ -511,11 +488,9 @@ class EvcsModelIT
 
       // Request free lots
       extEvData.sendExtMsg(new RequestEvcsFreeLots())
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(14400)
+      evService ! Activation(14400)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideEvcsFreeLots(
         // evB connected
@@ -523,7 +498,7 @@ class EvcsModelIT
         Map(evcsInputModel.getUuid -> int2Integer(1)).asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       // Request departing EVs
       extEvData.sendExtMsg(
@@ -531,33 +506,29 @@ class EvcsModelIT
           Map(evcsInputModel.getUuid -> List(evC.getUuid).asJava).asJava
         )
       )
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(14400)
+      evService ! Activation(14400)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideDepartingEvs(
         List[EvModel](evC.copyWith(20.0.asKiloWattHour)).asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       // Send (empty) arrivals in order to update next tick
       extEvData.provideArrivingEvs(
         Map.empty[UUID, java.util.List[EvModel]].asJava,
         Some(long2Long(18000)).toJava,
       )
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
       // waiting for ExtEvDataService
       scheduler.expectNoMessage()
 
-      serviceActivation ! Activation(14400)
+      evService ! Activation(14400)
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       evcsActivation ! Activation(14400)
 
@@ -570,11 +541,9 @@ class EvcsModelIT
 
       // Request free lots
       extEvData.sendExtMsg(new RequestEvcsFreeLots())
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(18000)
+      evService ! Activation(18000)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideEvcsFreeLots(
         // No EVs connected
@@ -582,7 +551,7 @@ class EvcsModelIT
         Map(evcsInputModel.getUuid -> int2Integer(2)).asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       // Request departing EVs
       extEvData.sendExtMsg(
@@ -590,33 +559,29 @@ class EvcsModelIT
           Map(evcsInputModel.getUuid -> List(evB.getUuid).asJava).asJava
         )
       )
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
-      serviceActivation ! Activation(18000)
+      evService ! Activation(18000)
 
       extEvData.receiveTriggerQueue.take() shouldBe new ProvideDepartingEvs(
         List[EvModel](evB.copyWith(10.0.asKiloWattHour)).asJava
       )
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       // Send (empty) arrivals in order to update next tick
       extEvData.provideArrivingEvs(
         Map.empty[UUID, java.util.List[EvModel]].asJava,
         None.toJava,
       )
-      extSimAdapter.expectMessage(
-        new ScheduleDataServiceMessage(adapterToExt.toClassic)
-      )
+      extSimAdapter.expectMessage(new ScheduleDataServiceMessage(evService))
 
       // waiting for ExtEvDataService
       scheduler.expectNoMessage()
 
-      serviceActivation ! Activation(18000)
+      evService ! Activation(18000)
 
-      scheduler.expectMessage(Completion(serviceActivation, None))
+      scheduler.expectMessage(Completion(evService, None))
 
       evcsActivation ! Activation(18000)